+++ CHANGELOG.md
@@ -202,6 +202,7 @@
 - Allow to select an external stock location on purchase order
 - Display origin and analytic template on move line grid
 - Improve Stock move report : add note and partner seq
+- Add an option to automatically confirm the bank orders generated from an invoice payment
 
 ## Bug Fixes
 - Check duplicate : Apply manual filter before duplicate check and allow to open form view.